requirements.txt
fastapi==0.70.0
uvicorn[standard]==0.17.0
psycopg2-binary==2.9.1
celery==5.1.2
redis==3.5.3
flower==1.0.0
alembic==1.6.5
SQLAlchemy==1.4.20
python-dotenv==0.19.2
